+++ source/RTSCamera.CommandSystem/src/Config/HotKey/CommandSystemGameKeyConfig.cs
@@ -1,0 +1,11 @@
+﻿using System.IO;
+using MissionSharedLibrary.Config;
+using MissionSharedLibrary.Config.HotKey;
+
+namespace RTSCamera.CommandSystem.Config.HotKey
+{
+    public class CommandSystemGameKeyConfig : GameKeyConfigBase<CommandSystemGameKeyConfig>
+    {
+        protected override string SaveName { get; } = Path.Combine(ConfigPath.ConfigDir, RTSCameraSubModule.ModuleId, nameof(CommandSystemGameKeyConfig) + ".xml");
+    }
+}