+++ source/RTSCamera/src/Config/HotKey/GameKeyConfig.cs
@@ -7,26 +7,6 @@ namespace RTSCamera.Config.HotKey
 {
     public class GameKeyConfig : GameKeyConfigBase<GameKeyConfig>
     {
-
-        private static GameKeyConfig CreateDefaultStatic()
-        {
-            return new GameKeyConfig();
-        }
-
-
-        protected override void CopyFrom(GameKeyConfig other)
-        {
-            base.CopyFrom(other);
-
-            ConfigVersion = other.ConfigVersion;
-        }
-
-        protected override void UpgradeToCurrentVersion()
-        {
-        }
-
-        protected override XmlSerializer Serializer => new XmlSerializer(typeof(GameKeyConfig));
-
         protected override string SaveName { get; } =
             Path.Combine(ConfigPath.ConfigDir, RTSCameraSubModule.ModuleId, nameof(GameKeyConfig) + ".xml");
         protected static string OldSavePathStatic { get; } = Path.Combine(ConfigPath.ConfigDir, RTSCameraSubModule.OldModuleId);